requirements.txt
Django==2.2.9
django-auth-ldap==4.0.0
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
ldap3==2.9.1
Markdown==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.6
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
